Updating to use default icon/conditions in not found cases
-- What
We have found that there are cases where an icon path can correspond
to two conditions at the same time, and these are not reflected in the
legacyMapping.

For the moment, if we don't find these keys, use a default icon and
conditions string. For now, this is nodata.svg and the
shortForecast (if present, otherwise "No data")