Reinstating normal obsKey usage, with updates
-- What
We noticed that some nws api icons have two simultaneous conditions as
a part of their url, and thus our legacyMapping keys and the
corresponding code that parses them out was built on incomplete
assumptions.

In this branch we temporarily had helper functions that would provide
defaults in these special cases.

After some discussion, we've decided instead to simply pick the
_first_ condition that can be parsed out of the icon url paths.

Note this required slight changes to the icon path parsing, to account
for the cases where there's an extra path component.