Updating config to use private Drupal folder for s3fs
-- What
The issue with upload/download is solved by configuring Drupal's
'private' file directory with s3fs. This allows us to successfully
upload, display, and download an image relative to the drupal
application's hostname/url.

The image will display on the node page but there is still an issue
getting it to display in the primary view.