Updating generation and spritesheet to remove class attrs
-- What
The presence of any single version of our SVG icons inline on the page
creates a reference to an `.st0` class, which then styles _any_ svg
child element that has that class set _anywhere in the document_.
This was happening with the spritesheet. We have removed all inline
classes on our spritesheet.