Updating daily forecast template with partials
-- What
We are now using a couple of helper partials in rendering the list
items for the daily forecast block. These help us iterate through the
different sections (today, detailed, and extended) in the main
template file without having to repeat ourselves.

Note that this commit does not yet include the implementation for the
"extended" periods. Forthcoming.