Adding stashed point and obs distance calc from that point
-- What
Now that we have rebased on point-based routing changes, we can stash
the given point as an instance variable on the data service object.

As a consequence, we should calculate the distance to the observation
station from that reference point if it is available.

Calculating from the grid is still an option if no reference point is available.