Adding cases of null ends and/or expires
-- What
For now, if a given alert is missing the ends field, we try the
expires field as the end time. If that is also missing, we simply
ignore the alert for the purposes of displaying it in the Hourly
Table.

In the future, we might want to slightly alter this assumption. For
example, we might want to extend the endTime to then be the end time
of the last period in the hourly forecast, so the alert extends across
the full time.