Adding some e2e tests with caveats
-- What
Cypress seems to operate non-deterministically (or with flaky
behavior) when it comes to issuing keyboard events and clicks on the
elements in this component. Sometimes the simple test cases pass, and
sometimes they don't.

Additionally, the click and double enter actions, which should trigger
a form submission and page navigation, aren't nagivating at all, even
though they do in a live environment.