Fixing how we filter future day periods in Daily trait
-- What
Previously, we were setting the $now value _after_ we had already
attempted to filter future daily periods. This prevents us from having
any injection of a different $now value for testing purposes and
having it filter appropriately. It was also not mapping to the correct
timezone, which we have changed.