Fixing loss of focus when clicking/tapping list item
-- What
Some browsers will attempt to fire a focus event on a clicked/tapped
element _before_ dealing with any click handlers on those
elements. This is because the focus event will be triggered during
mousedown time, not click time. The fix here is to add a
preventDefault() call to stop normal browser behavior before a full
click event is fired on any of our list items.