Removing unneeded function
-- What
Greg highlighted a simpler way to determine whether or not to only
show the low temperature: if there is only 1 'today' period, then it
is a night/overnight period and thus should only show the low.

Therefore we can remove the more complicated time calc function, since
we already have the information.