Updating bundles for AFD products
-- What
We now trim the AFD references for each WFO (in this case OKX and TAE)
down to just 10 entries. We also then add each of product data files
that are referenced. This last step is necessary because otherwise
tests or examples would reach the live API to get the product by
ID. The problem is that after some time these product IDs can no
longer be found by the API. These changes will help prevent that inconsistency