Fixing for PoP edge case in gridpoints endpoint
-- What
In our test data, there are gaps in the gridpoint PoP hourly
information. These might not exist in real life, but we should check
for them when trying to compute a max PoP for a given period/day.

If there is no `probabilityOfPrecipitation` for a given hour, we
return 0 as the probability when trying to compute the max for the period.