Fixing list item / tabpanel a11y concern
-- What
It turns out that you cannot set role="tabpanel" of a li element. This
changes the element's semantic meaning, which axe thankfully
caught (this is not well documente din the WCAG for tabs
specifically).

We have to introduce an intermediate div and adjust the DailyForecast
component to use that div as the tabpanel.